Add total-risk-score chart
This MR adds a new chart component which will
render the "Total Risk Score" for the security
dashboard.

The chart is a gauge chart, with a colour-based
meter to visualise the calculated risk of the
current project or group.

Besides adding a new Vue component,
it also adds a storybook entry,
which can be used to discover and test the
component.