Remove custom exit codes for monitoring
We tried those as an experiment to be able to tell
which failure category a given CI/CD job failed with.

We now have logic for that purpose outside of the CI.

This commit removes the previous logic, as it was
anyways disabled.

We only keep the exit code 111 for "disk almost full"
to auto-retry the job, as it would have most probably
failed anyways.