Dependency Graph creation with path caching
This commit introduces path caching for building the dependency graph.

Prior to this change, for large graphs or those with decent levels of
interconnectedness, we would traverse the same subgraphs multiple times
to find the paths between a top_level_component and all components in
that subtree. We would traverse from the top_level_components downwards.

We now traverse upwards from the leaf components to find
top_level_components, and when we do find a top_level_component we
cache the distance between that top_level_component and every path
partial. As we continue traversing, if we revisit a node we have seen
before we can use the cache as a shortcut to find the distance to all
top_level_components above this node, cutting out significant rework.

Changelog: changed
EE: true