Set GOLANG_FIPS to 1 to test in FIPS mode
While we build Workhorse with a FIPS compiler, the tests aren't run in
FIPS mode unless a FIPS kernel is used, or `GOLANG_FIPS=1` is set in the
environment. We would have caught TLS issues in the `golang-fips`
upgrade had we been in this mode.

Let's test FIPS by setting this in the environment.

Relates to https://gitlab.com/gitlab-org/build/team-tasks/-/issues/13