Fix ES sync with deleted vulnerabilities
By the time we call sync_elasticsearch method in
Vulnerabilities::Removal::RemoveFromProjectService,
the vulnerabilities have already been deleted. So when we try to
sync with elasticsearch we end up passing an empty array
to the BulkEsOperationService. This MR fixes that
by loading the records in memory before syncing.

Changelog: fixed
EE: true