Fix member_role_policy ACL template for Openbao
Prior to this change, we had an issue with the
member_role_policy ACL template not
handling the case where member_role_id is nil in
the JWT payload.

This change fixes the issue by updating the
member_role_policy template to properly handle the case where member_role_id is nil.

Changelog: fixed
EE: true