Add not null constraint to deploy token project_id and group_id
The `BackfillDeployTokensShardingKey` background migration set the
sharding key for all existing deploy_token records in 17.4. However,
since no constraint was added at the time the database has not prevented
invalid records being created, and we have found that invalid records
exist in staging from a test restore of an older deleted project.
No invalid records exist in production.

These migrations re-execute the same code used in the background
migration to initially populate these columns, and add a not null
constraint to satisfy the requirements for the sharding key.

Changelog: other