Support agent inputs
Make all agents declare the same 3 inputs:

- `context:{previous_agent_id}.final_answer`: The previous agent's final
  response
- `conversation_history:{previous_node_id}`: The previous agent's
  message history
- `context:goal`: Workflow goal

With aliases, that can be used in their goal templates.

It also refactors away the referring of the `agents` collection directly
and instead all accessing of agents is through `nodes` which is a
combination of agent + step.

This is because the same agents may be used multiple times within the
same flow, and we want to accurately refer to specific agent+step
combinations when identifying them, determining the previous one, etc.

https://gitlab.com/gitlab-org/gitlab/-/issues/560719