IncrementalIndexing must record the correct last queued item correctly
Previously, the IncrementalIndexingService updates the `initial_indexing_last_queued_item`
value. This should not be the case since that value is specific to initial indexing
and used for marking a repository as ready.

Here, we change the logic so that IncrementalIndexingService updates a different value:
`incremental_indexing_last_queued_item`.