Fix: Atom feed broken when supplying ref_type
Contributes to https://gitlab.com/gitlab-org/gitlab/-/issues/562693

**Problem**

The atom feed for commits fails to render entries when the `ref_type`
parameter is provided. This occurs because the qualified ref name has a
suffix `.atom` that is not removed.

**Solution**

Recalculate `@fully_qualified_ref` after `.atom` suffix was removed.

Changelog: fixed