Fix moving epics to a new group
When moving an epic to another group, we copy all the related data over
to a new epic, or rather its related work item.

As part of the epic work item migration, every epic has a correlating
work item (issue_id), and for some shared associations between the epics
and the issues table, we still use the old epic_id OR the new issue_id.

Usually this is handled via a `type` column, like `noteable_type` and
`notable_id`. In the case of `description_versions` we either set the
`epic_id` or the `issue_id` in the column, and have a XOR NOT NULL
constraint.

So there are still many description_version records with an `epic_id`,
and when we copy over the data as part of moving the epic, we copied
over the `epic_id` AND the new work item `issue_id`. This resulted in a
database constraint check error, since we can't set both values at once.

This change fixes the copy mechanism by removing the `epic_id` when
copying the description_version attributes. The `epic_id` is no longer
needed, as we can associate it anyway with the new work item
(`issue_id`).

Changelog: fixed
EE: true