Generate a user-friendly name for workspaces
Workspaces name is generated by SecureRandom.alphanumeric so far. This
does not give a user-friendly and memorable name. This commit changes it
such that the workspace suffix is created by the combination of a fruit,
an animal and a color, downcased, parameterized and joined by a hyphen.

Changelog: changed
EE: true