Disable gdk-update job in stable branch pipelines
This disables the `gdk-update` job in stable branches because it always
uses the latest GDK image, which might have become incompatible with the
stable branch.

The goal of the `gdk-update` job is also to prevent impact to active
development by shielding engineers from failures due to e.g. faulty
migrations. This is not that relevant for stable branches, especially
with the upcoming introduction of GDK database sandboxing.