Only create ToC for headings with links
If a heading in a file exists without a link inside it,
creating items for the table of contents will through
an error. A filter on headers ensures that only those
elements that do have an ancore tag are added to the array.
Affecting rich viewer render of AsciiDoc, reStructuredText.

Changelog: fixed