Always build both Tailwind CSS bundles
This ensures that both `tailwind.css` and `tailwind_cqs.css` are built
when running the GDK. This meands that the
`USE_TAILWIND_CONTAINER_QUERIES` doesn't need to be set manually anymore
as both bundles are guaranteed to be available.
This makes it easier to toggle between both states of the
`tailwind_container_queries` feature flag.