Reimplement the refreshCurrentPage JS functionality
The previous implementation relied on `window.location.assign`. If the
URL contained a hash it caused to navigate to the specified portion,
instead of reloading the page.

`window.location.reload` is a more robust functionality that causes the
browser to refresh regarding if the URL contains a hash or not.

Changelog: fixed