Fix escape characters in the nodejs coverage regular expression
The regular expression for the nodejs coverage contains unescaped pipe '|' characters when
rendered to the page. This commit adds a backslash to the rendered page so that the regex
example in the documentation pulls out the correct coverage value.

If you used the original, unescaped version of the regex, the coverage would tend to pick out "201" 
from the artefact upload success messages, giving the code an incorrect coverage score of 201%