Fix webauthn authentication in Firefox
In Firefox, passing a reactive prop (Proxy) as `publicKey` in
navigator.credentials.get` causes an internal error: `Promise rejection
value is a non-unwrappable cross-compartment wrapper.` The most probable
reason for this is that `convertGetParams` returns some sophisticated
data structures (ArrayBuffer and Uint8Array).

We fixed this by passing to `publicKey a non-reactive prop in this way:
`publicKey: convertGetParams(this.webauthnParams)`.

Changelog: fixed