Refactor tests abstract pipeline details workload users
These tests are testing classes that only depend on the `Ci::Workload`
class. The `Pipeline` is an internal implementation detail of the
`Ci::Workload` class and we may want to change that in future. To make
it easier to refactor this we want to avoid referencing pipelines
altogether in code paths that use workloads (except inside the actual
Workload classes).

This just makes some trivial refactors to tests that align with this
objective.