Fix External::Context spec timeouts
We have consistently experienced timeouts in CI (only CI) from two
specs that both rely on External::Context. Determining the exact cause
has been challenging, but it seems that passing `project: nil` to
`External::Context` can result in threading issues that cause the specs
to hang.

There's always a project present in production, so this commit updates
the specs to pass a project to Context. Hopefully that resolves the
intermittent timeouts.

https://gitlab.com/gitlab-org/gitlab/-/issues/560096
https://gitlab.com/gitlab-org/gitlab/-/issues/558142