Convert Tailwind script to CommonJS
This rewrites the script and changes from
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/202799 into
CommonJS, so that the module cache can be cleared.

For now, there's no way to clear the equivalent cache for ECMAScript
modules: https://github.com/nodejs/node/issues/49442.

An alternative approach could be to use `node:worker_threads`, but this
is the simpler approach.