Fully rollout (remove) feature flag for cloudflare for Duo Workflow
In In https://gitlab.com/gitlab-org/gitlab/-/issues/509586 we introduced
a feature flag for determining whether or not to send Duo Workflow
requests via Cloudflare. This was enabled per
https://gitlab.com/gitlab-org/gitlab/-/issues/509586#note_2714079014 and
it seems we do not have any new networking issues.

Now we want to fully roll this out and remove the feature flag so that
self-managed customers start using cloudflare.

The only change in this MR is in the `default_service_url` method. All
it does is return the cloudflare URL now instead of the direct URL for
the service. Cloudflare then proxies to this same underlying service
URL.

Changelog: changed