Expose raw item version name in GraphQL
This change renames the current `versionName` field to be
`humanVersionName`, as it calls a method on the backend that returns a
version prefixed with `v` (and suffixed with `-draft` if in draft).

The `versionName` field now returns the raw version name that the
backend uses.

The motivation is because this is the format that versions must be
passed to the backend as arguments.

The queries and mutations are experiments, which is why this MR has no
changelog.