Explicitly define the MR and Issue create button text
Please see: https://gitlab.com/gitlab-org/gitlab/-/issues/543136 for more information.

Note these:

- "Rails' automatic conversion of class names (model_name.human.downcase) which is not working well for non-English languages".
- "does not allow translators to provide appropriate grammatical constructions for their languages"

While dynamic humanization and downcasing of the model name is graceful,
it's probably not worth the localization harms caused, and the potential
overhead of explicitly defining the strings for each type of issuable is
not going to be all that challenging (a small handful of items, perhaps).

At the moment I believe this is exclusively "merge request" and "issue".

Changelog: fixed