Update gitlab secrets manager payload to runner
Before this change, we were sending the
Gitlab Secrets Manager payload to the Runner
masquerading as a Vault secret. Now, we have implemented
the functionality in runner to handle the Gitlab Secrets Manager payload
Now, we need to update the payload to match the new format

This commit updates the Gitlab Secrets Manager
payload structure to align with the changes made in the Runner.

https://gitlab.com/gitlab-org/gitlab/-/issues/561479

Changelog: changed
EE: true