Attempt to fix flaky specs for protected branches
We've been experiencing flakiness in the protected branches specs for a
while. I'm not convinced I'm solving it here but it looks incorrect.

We have a path that redirects to the repository settings now. If you
load this path manually it shows the protected branches in collapsed
view. I suspect that when we click "Add protected branch" while the
section is collapsed it doesn't work and then when we try to add an
allowed to push value it waits until the timeout because the dropdown is
invisible.

I'm not convinced this is the problem as the test is flaky. I would have
expected this test to always fail if this is what's happening but
perhaps the section sometimes shows and sometimes doesn't which causes
the flakiness.

We can test this approach and if we still have flakiness then we'll need
to reevaluate.