Vulnerability merge_requests field
Vulnerabilities support multiple linked merge requests, but
the API only exposes one merge request. When there are
multiple added, the merge request exposed is chosen in a
non-deterministic way.

This commit adds a `merge_requests` field to the graphql
vulnerability type.

It also forces `merge_request` to be the oldest MR in the list.

Changelog: added
EE: true