Make FileLocationType.endLine nullable
This graphql schema is not aligned with our input validation, which
expects that this field may be nullable. A file location with no endLine
indicates a one-line issue. The frontend already expects that this field
is nullable. This fixes a bug where the graphql query may error if there
if a vulnerability that has file location details with no end line.

Changelog: fixed
EE: true