Add details in schema checker inconsistency object
This is needed to provide richer detail as to what is not right with the
database. For example, with sequence ownership we want to display the
current owner and the expected owner.

This commit bumps the schema checker version from v1 to v2 accordingly.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/396878

Changelog: added