Add migration to make some foreign key names consistent
The database diagnostics page identified a number of missing foreign
keys, but they actually exist in other names. It appears that these
inconsistencies occurred for a number of reasons:

1. ActiveRecord automatically generates foreign keys in the form of
`fk_rails_<hash>`, but during the squashing of migrations these foreign
keys got renamed to `fk_<hash>`. This commit adds a migration to rename
known foreign keys back to `fk_rails_<hash>` again.

2. Some indexes were not renamed when their respective tables were
renamed. `<hash>` is calculated with the original table name. This
commit adds a migration to rename the foreign keys using the hash with
the new table name.