Implements link maven upstream form
Render form when Link existing upstream button
is clicked.

Use TestUpstreamButton component to implement
test upstream button.

Show list of upstreams not part of registry in
a collapsible listbox.

When upstream is selected fetch upstream details
and show summary.

Linking the form submit with the API will
be implemented in a follow-up MR.

EE: true