Add target project claims to CI ID Tokens
Recently, we updated the logic to set the project
related claims in the CI ID Tokens for the case of
MR pipelines from forked projects which run inside
the target project. However, this caused an issue
where there exists no claims that can unambiguously
identify the project where the pipeline is running.
For usecases, where we want to allow ID tokens for all pipelines
running within a project even from forks, we need to ensure that
the target project claims are included in the token.

This change fixes the issue by adding the
job project claims to the CI ID Tokens.

[Issue](https://gitlab.com/gitlab-org/gitlab/-/issues/563038)

Changelog: added