Assign project target_type to newly created project events
Previously, the `target_type` for events was not set for project-related
events. This causes a bug where using the Events API
and filterting for `target_type=project` always returns
an empty result.

This fixes that by setting the `target_type` to `project`
for relevant events.

Note that existing records will be backfilled in a later step.

Changelog: fixed