Fix Rake test failing with single-db-ci-connection
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/202736 was failing
in the `single-db-ci-connection` tests because the
`skip_if_multiple_databases_not_setup` was not checking
the right configuration. In the `single-db-ci-connection`
case:

1. The `main` and `ci` databases are configured separately
but share the same database.
2. In the `ci` database config, `database_tasks` is set to `false`,
which disables database-specific Rake tasks.

Fix this failing test with a new helper method
`skip_if_database_specific_tasks_not_available` that considers
`database_tasks`.