Render unverified tag signature instead of 500 page for timeouts
When the RPC for tag signatures exceeds the timeout duration we
display a 500 page. Timeouts appear to be uncommon and only for
larger repositories.

In order to facilitate the rollout of the
tag signatures features that are still behind the
`render_gpg_signed_tags_verification_status` and
`render_ssh_signed_tags_verification_status` feature flags without
disrupting the tags views for larger repositories that would be
affected by timeouts we recover by using the existing behaviou of
displaying an unverified badge.