Fix generated ref commit association
This fixes a deprecation in the association definition:

DEPRECATION WARNING: Setting `query_constraints:` option on
`MergeRequests::GeneratedRefCommit.belongs_to :merge_request` is
deprecated. To maintain current behavior, use the `foreign_key` option
instead.

We need to change this as this was causing a spec failure in
`ee/spec/models/merge_trains/car_spec.rb` when we save the associated
generated ref commit records.