Avoid using discarded connection pools in load balancer
For reasons we don't fully understand yet, it appears that it is
possible at startup for Puma to get into a state where it repeatedly
fails to boot with the error:

```
ERROR: The hook worker_start_hooks failed with exception
(NoMethodError) "undefined method `delete' for nil:NilClass".
```

It appears that this happens when a child Puma process attempts to use
the master's discarded connection pool at startup. Here is what appears
to be happening:

1. The master process starts and initializes a load balancer.
2. The master process forks.
3. As a result of the fork, Rails' `ForkTracker` calls `discard!` on all
   connection pools, including the one used by the load balancer.
4. A child process sets up its own load balancer and initializes a
  `Gitlab::Database::LoadBalancing::Host` for each of the configured
  replicas.
5. `Host` attempts to create a Prometheus gauge but first checks whether
   the application settings have Prometheus metrics enabled.
6. To perform this check, the load balancer created in the master
   process is used because the child process is still in the process of
   setting up its own load balancer.
7. The load balancer calls `Host#refresh_status` to check the status
   of the replicas.
8. Because `ForkTracker` already discarded the connection pools, the
   load balancer hits the `undefined method` after releasing the
   connection.

For this to happen, a number of things must have occurred:

1. The memoization for `prometheus_metrics_enabled?` had to be cleared
   for the child process to make another database query. The master
   process normally sets up this memoization in the
   `7_prometheus_metrics.rb` initializer.

2. Normally the load balancer trusts that the host objects have been
   online for the first `replica_check_interval` (default 60 s), but
   this suggests it took longer than 60 seconds between the
   initialization of `HostList` and the read attempt.

3. The process memory store that caches `ApplicationSettings` has expired.

While it appears difficult to reproduce this behavior, we can
artificially create those conditions by modifying the code and settings
to simulate the conditions.  The fix is to ensure that discarded pools
are never used, which allows the application to start properly.

Relates to https://gitlab.com/gitlab-com/request-for-help/-/issues/3317

Changelog: fixed