Fix component lookup by partial semantic version
`~latest` already is documented as requiting the catalog, but _partial
semantic version_ also requires the catalog for lookup. Currently the
lookup fails and the generic error message is returned:
> Component 'gitlab.com/group/project/template@1' - content not found

1. Add `invalid_usage_of_partial_semver` similar to
   `invalid_usage_for_latest` checking for partial semantic versions.

2. Introduce a separate error message for partial semver similar to
   latest, which shoved the user into the right direction, that partial
   semver also requires the component to be published to the catalog.

3. Add test cases for partial semantic version.

4. Update documentation.

Apply suggestion from @lyspin and @Kasia_Misirli

change `available` to `published` to make it clear, that a _available
tag_ is not enough, but that it must be published to the catalog to be
available.

Shortened section about selection by partial semantic version.

Signed-off-by: Philipp Hahn <p.hahn@avm.de>