Backfill workspace_agentk_states for all workspaces
This migration will create desired_config for all the workspaces
regardless of their state and create an entry in the table
`workspace_agentk_states`. This is required as part of freezing
desired_config of Workspaces in the create step instead of
reconciliation.

Any workspace that fails migration will be marked as `Terminated`.
Failure details will be available in the logs and in the
`workspace_agentk_states` table. Cluster administrators are required to
clean up any orphaned workspaces.

Changelog: other
EE: true