Disable backfill_workspace_agentk_states migration
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/199102
added migration logic to backfill workspace_agentk_states table for
all the existing workspace. This logic is slightly incorrect. The jsonb
which is supposed to be stored inside this table is supposed to have an
array and we are saving an object. In other words, root level node is
not required in this table.