Make private mean private to a project
Currently, our implementation of private flows and agents means a flow
or agent is available to be used in all other projects any developers+
of the project is a maintainer of. This includes projects that are
outside the top-level heirachy.

This currently allows a leaky system where a `gitlab-org/gitlab`
private flow is able to be used in `atlassian/jira` because a developer+
of `gitlab-org/gitlab` is a maintainer of `atlassian/jira`.

There are some problems:

- The project owners have no actual ability to determine the privacy of
  the flow or agent. It's instead determined by what other projects
  their members are maintainers of, which can change and expand at any
  time.
- It's confusing and an unexpected result
- A project cannot make something properly private to itself

This change allow customers to:

- Create a flow or agent for private use within their projects
- Know exactly the extent of what privacy means, based on group and
  project hierarchy and not individuals within them

In future, for group-level flows
(https://gitlab.com/gitlab-org/gitlab/-/issues/553912), a private
group-level flow or agent would be for use within only the descendant
projects.

https://gitlab.com/gitlab-org/gitlab/-/issues/557250