Fix security widget polling indefinitely when there are sboms
This bug happens then there is an sbom that does not produce
dependency_scanning results, and no other dependency scanning report in
the pipeline. We have a redis optimization to reduce the number of
database queries executing when polling. It blocks the database checks
from executing until a redis key is set for the report. That key is
being set if an SBoM produces results, but if it doesn't then the report
is shown as parsing indefinitely.

Changelog: fixed
EE: true