Add an init method to prepare for Puma 7
With https://github.com/puma/puma/pull/3297, Puma 7 raises an exception
if Puma's options are accessed before they are fully loaded. Currently
we pass these options during loading of the Puma config file, which then
triggers this exception. These options are used to determine whether
Puma is in clustered mode by checking the number of workers.

To prepare for Puma 7 and avoid this exception, add a
`set_puma_worker_count` method that registers the number of workers
directly.