Save execution config to CI Job Definition
As a part of the CI storage optimization effort, we started to save
the runtime job configurations into the `p_ci_job_definitions` table.

In this change, we are starting to save execution configs (run steps)
to the table too. This change only covers the saving this data to the
table. We are actually duplicating the data as we also save this data
to the `p_ci_builds_execution_configs` table. However, our plan is to
remove that old logic in the future.