Call update_user_group_member_roles with correct args
Currently, when SAML groups are synced, the users'
user_group_member_roles records are only correctly updated when they're
not yet a member of the group. However, when existing memberships are
updated during sync, the user_group_member_roles records are not
updated.

Here, we update Members::Groups::CreatorService used in
Groups::SyncService to correctly call update_user_group_member_roles on
the created/updated Member record. An old_values_map hash argument is
now passed in during the call so that the ::Authz::UserGroupMemberRoles::UpdateForGroupMemberService knows if the
member was updated and update the user's user_group_member_roles
accordingly.