Add more logging to workhorse duo agent gRPC logic
We found in
https://gitlab.com/groups/gitlab-org/-/epics/19253#note_2735093445
that there are cases where we send an HTTP request to GitLab and then it
seems to never get back to Duo Workflow Service and this logging may
help us figure out how far we got before disconnecting.

We probably will require more logs but this is a start because knowing
whether or not we get to this point and whether or not there was an
issue might already be informative.