Set fixed-height on sticky header to prevent layout thrashing
The sticky header being deployed at the medium breakpoint
was "jumping" at a high rate if users tried to scroll while
a new merge request was being created. This was deemed a layout
bug and could be problematic for users with vestibular or
photosenstivity issues.

The root cause of the issue was determined to be a container
type scroll-state applied to a container with height: auto.
Containers that do not have enough space for their content
to expand or reflow can cause this visual jump.

Changelog: fixed