Improve Media Bubble Menu accessible labels
The media_bubble_menu uses a div[role="group"]
to contain two label + input groups. This was
identified as a possible accessibility issue
because the inputs' native roles might be
ignored. MR adds an accessible container label
and an autofocus on the first input after
confirming the native roles were announced
properly to the big three screen readers.

Changelog: changed