Drop NOT NULL on mrdf partition.new_path
Due to a mistake in the sequencing of migrations, we dropped the NOT
NULL constraint on `merge_request_diff_files` AFTER we create the
partitioned table. This is due to the migrations being run in their
timestamp order.. so even though the NOT NULL constrain was merged
first, it runs second. This didn't show up on local dev because we had
mixed-state, and weren't testing the _interaction_ of these 2
branches.