Stop generating Ruby coverage HTML in merge requests
This stops generating the Ruby SimpleCov coverage information in HTML on
merge requests.

The main reason for this is to improve performance. Generating this
information is computationally expensive and the resulting HTML is
barely viewable at the size of the GitLab code base because it is almost
400 MB large and contains all of GitLab's source code.

The `rspec:coverage` job is on the critical path of almost all
pipelines, so reducing its size will almost 1:1 translate to a reduction
in overall pipeline duration.