Add spdx or custom_license constraint to software_license_policies
This MR adds a database constraint to match the model
validation. This constraint could not be added before
because the software_license_policies in the sec
database was not truncated yet.

Changelog: added

This MR adds a database constraint to match the model
validation. This constraint could not be added before
because the software_license_policies in the sec
database was not truncated yet.

Changelog: added