Sort built in tools by name
The built in tools are a Hash of data.

This change raises the `max_page_size` to `1_000` as returning this data
does not involve any DB requests, and currently the frontend requires
the complete data set in order to do filtering on the client side.

It returns the list sorted by name.