Add name filter to workItemAllowedStatuses field
Add a name filter argument to the
workItemAllowedStatuses GraphQL field to allow
filtering work item allowed statuses by name.

The implementation is a placeholder for now,
with the actual filtering logic to be implemented
in a follow-up MR.

The name argument is optional and will enable
frontend consumers to filter the list of statuses
when needed, improving the user experience
in the status filter dropdown.

Changelog: added
EE: true