Link retried job to environment before the job is started
Deployment records are created on demand when a job is due to start,
and this is when the deployment ID is set on the job's job_environment
record. If the job starts before it has been linked to the environment,
the deployment ID is never set because the link does not yet exist.

This changes the order to always create the environment link first,
followed by starting the job (and linking the deployment to the
environment).