Add NOT NULL constraint for identities -> user_id
Ensures a user_id is always set for the identities table. This is
required for upcoming sharding changes. The application already enforces
that identities are created and updated with a user_id, but so far we
have not enforced that with a database constraint.

Since identities without a user_id cannot be used for any purpose, we
add a migration to remove any such corrupted records in case they exist
on a self-managed instance.

Changelog: changed