Clean redis state between test ghost user tests
This tests code which relies on `Gitlab::GithubImport.ghost_user_id`.

This value is cached in redis, and so tests which destroy and re-create
the ghost user can create a mismatch between the ID in the database and
the ID stored in the cache.

Example failing case on `master`:

```bash
rspec spec/lib/gitlab/github_import/importer/releases_importer_spec.rb \
  spec/lib/gitlab/legacy_github_import/user_formatter_spec.rb \
  --seed 40225
````

Clearing the redis state for these specs helps make sure we're using the
correct value.