UI for display token status on Pipeline > Security Details Page
Creates the display for the Pipeline > Security Details page
without connecting the back end yet. Adds an experiment badge
to the token status section. Due to refresh logic not existing
for Pipeline > Security Details we need to ensure that it does
not get rendered within ValidityCheck component and can instead
use the TokenStatusBadge to display the token status in this
case.

Changelog: added
EE: true