Apply `with-gl-container-queries` class based on FF
This updates the logic for applying the `.with-gl-container-queries`
class to only check for the `tailwind_container_queries` feature flag's
state rather than the whole UI redesign flag. This lets us migrate
responsive mixins to the new ones we built in GitLab UI at the same pace
as the CSS utils migration.