Create work_item_descriptions table partitioned by namespace_id
This recreates the previous MR for adding the work_item_descriptions
table, but this time partitioned by namespace_id instead of
root_namespace_id.

The reason for the change were problems we ran into when syncing the
issues table to the work_item_descriptions table when a namespace
transfers to another root namespaec. In this race condition, the
work_item_descriptions table would have had a record in the old and in
the new partition.

Since the namespace_id is immutable, we can partition by it without this
problem.

Changelog: added