Add AuthenticationEventArchivedRecord table
Add an authentication_event_archived_record table and bare-bones model.

The purpose of this table is to store deleted rows from the operational
`authentication_events` table. We will be archiving records from
`authentication_events` and will store deleted records in the new table
as a soft-delete mechanism.

The archive table is meant to be temporary and we will remove it after
the retention policy has been enforced on `authentication_events` and
is verified.

Changelog: added