Fixes target projects endpoint 404 on compare revisions view
Fixes the target projects endpoint raising a 404 when viewing the
compare revisions page when you do not have permission to create a merge
request. This was caused by the endpoint being under the merge request
creations controller which checks permissions for creating a merge
request. As this endpoint is shared between 2 views this change moves
the endpoint to a controller that will render a 404 when the user
doesn't have the right create permissions.