Remove health signal check for MR Webhook Worker
When we defer the worker on the database health check, we can get a very
large number of deferrals which can cause a spike of workers being
executed at once, leading to errors in Prometheus for the worker.

This change removes the merge_requests table from the health signal
check to prevent these deferrals while keeping the basic health check
structure in place.